Rename ingredient modals for clarity
